Changed invoice supporting documents to include data for all bills.
Updated the basic BopsScan class with a more honest "DocScanned" name instead of LoadRef, changed the underlying data layer query model to retrieve more kinds of docs, and reworked the dialog that displays the supporting documents to use the new features. Added unit tests to teh SCM Ops Core solution and revved the version information for the RFS Invoice Utility in the process.