Fixed a minor issue preventing storage billing rollback from working.
Julie C. texted me this morning with the need to roll back storage billing 145, and by my memory that was a function that couldn't work. I had to dig around for quite a while until I discovered the issue: during the data layer conversion, I missed converting a List to a Dictionary. So every attempt to look up an inventory item's tracking details was failing--a problem that looked much worse than it was. A few lines of code later, the issue was solved.