Fixed an issue with storage billing.
There was a single place in the code where a storage billing unit conversion could allow for negative revenue to be calculated in the event that an unsupported unit was specified. The routine should now throw an exception in such cases. Support was added for metric tons as well, which was the offending unit that evinced the bug.