Removed Budway logo and improved the login dialog.
Dan H. asked me to remove the Budway logo entirely from the main form and not replace it with anything. I deleted it and moved the investigation controls around to take advantage of the screen real estate. Also improved the login dialog by fixing the tab order and making the OK button the default for the form.