Added support to record Windows user IDs who create/finalize invoices.
Dan H. told me yesterday that he's being asked to report on which users invoice the most revenue. We don't have a real authentication piece in the invoice utility, but he said he could settle for the Windows user involved. I've added fields to the database and added code to set the fields in the proper places. I've also added a new context menu item and windows form to see those user details. This brings the utility to v1.2.0.