Updated all the Crystal Reports files to work with SQL16.
This is a painful commit. The smart thing to do would be to store a connection string in the app config file, parse out its important bits, and programmatically set them into the Crystal Reports report objects at run time. But the only ways I've found to do that take MINUTES for each report to open every time. That's ridiculous. So for the time being, I'm going with hard-coded values in the report files and hard-coded data in the C# code, which still has to shove the password into the report object at run-time to avoid having a dialog pop up to the user--every single time.