Fixed RFS Invoice Utility sorting issue when with manual bill invoices.
Julie C. demonstrated to me how the invoice shipper and bill-to clients were not sorted in any meaningful order on the form to create an invoice from a single, manual bill. I looked into it and found the data was being added to the fields unsorted and relying on the combo-boxes to do that, only they weren't set to sort. Now they're set to sort.