Fixed a long-standing bug when calculating warehouse results.
I can't believe this bug has remained in here since the days when we first started dealing with warehouse services. The short version of the story is that the code that loaded the warehouse documents did NOT properly filter on the basis of their type (receiver vs. shipper/BOL), so recent receiver IDs were causing BOLs from years ago to be recalculated, fail, and be mis-reported as meaningful errors.