Fixed a problem with merging invoice supporting documents.
When the code was created, we decided that the only documents we'd ever be handling were .pdf files. Unsurprisingly, folks started including .jpeg files which the PDF merger couldn't handle. This new version of the RFS Invoice Utility changes the code to include only .pdf files when merging all invoice supporting documents.